+++ src/Rule/Performance/CheckCallsInConditionsRule.php
@@ -22,8 +22,8 @@ use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\StaticCall;
 use PhpParser\Node\Stmt\If_;
 use PHPStan\Analyser\Scope;
+use PHPStan\Rules\IdentifierRuleError;
 use PHPStan\Rules\Rule;
-use PHPStan\Rules\RuleError;
 use PHPStan\Rules\RuleErrorBuilder;
 use PHPStan\Type\ObjectType;
 use PHPStan\Type\VerbosityLevel;
@@ -85,7 +85,7 @@ final class CheckCallsInConditionsRule implements Rule
 
     /**
      * @param If_ $node
-     * @return RuleError[]
+     * @return list<IdentifierRuleError>
      */
     public function processNode(Node $node, Scope $scope): array
     {
@@ -97,7 +97,7 @@ final class CheckCallsInConditionsRule implements Rule
     }
 
     /**
-     * @return RuleError[]
+     * @return list<IdentifierRuleError>
      */
     private function processExpr(Expr $expr, Scope $scope): array
     {
@@ -114,6 +114,7 @@ final class CheckCallsInConditionsRule implements Rule
 
             foreach ($slowCallsInPreviousParts as $slowCall) {
                 $errors[] = RuleErrorBuilder::message('Performance: "' . $slowCall . '()" is called in condition before expressions which seem to be faster.')
+                    ->identifier('efabrica.slowCallsInCondition')
                     ->tip('Move faster expressions to the beginning of the condition and calls to the end.')
                     ->line($expr->getLine())
                     ->build();