+++ src/Rule/Schema/UnusedProperties.php
@@ -48,9 +48,6 @@ final class UnusedProperties implements Rule
 
     public function processNode(Node $node, Scope $scope): array
     {
-        if (!$node instanceof CollectedDataNode) {
-            return [];
-        }
         $schemaUsage = $node->get(SchemaUsage::class);
         $this->schemaDefinitions = $this->convertSchemaDefinitions($node->get(SchemaDefinitions::class));
 
@@ -65,7 +62,10 @@ final class UnusedProperties implements Rule
                     $schemaName,
                     implode(',', $unusedProperties),
                 ))
-                    ->file($this->schemaDefinitions[trim($schemaName, '\\')]['file'])->line($this->schemaDefinitions[trim($schemaName, '\\')][3])->build();
+                    ->identifier('efabrica.schemaUnusedProperties')
+                    ->file($this->schemaDefinitions[trim($schemaName, '\\')]['file'])
+                    ->line($this->schemaDefinitions[trim($schemaName, '\\')][3])
+                    ->build();
             }
         }
 
@@ -94,12 +94,11 @@ final class UnusedProperties implements Rule
         $result = [];
         foreach ($schemaDefinitions as $key => $value) {
             $tmp = $value[0];
+            /** @var array<int, array{key: int, name: string, type: string}>|null $attributes */
             $attributes = json_decode($tmp[2], true);
             $tmp['attributes'] = [];
-            if (is_array($attributes)) {
-                foreach ($attributes as $attribute) {
-                    $tmp['attributes'][(int) $attribute['key']] = (string) $attribute['name'];
-                }
+            foreach ($attributes ?? [] as $attribute) {
+                $tmp['attributes'][$attribute['key']] = $attribute['name'];
             }
 
             $tmp['file'] = $key;
@@ -190,10 +189,6 @@ final class UnusedProperties implements Rule
             }
         }
         foreach ($attributesArray as $attributes) {
-            if (!is_array($attributes)) {
-                continue;
-            }
-
             foreach ($attributes as $attribute) {
                 if (isset($result[$attribute['key']]) && $result[$attribute['key']] === false) {
                     continue;